requirements.txt
fastapi
uvicorn
SQLAlchemy
psycopg2-binary
pytest
requests
httpx
alembic
python-dotenv
trafilatura
playwright
apscheduler
BeautifulSoup4
crawl4ai
transformers
torch
tensorflow
unstructured